Fade only the dark keypad, not the live one
This cuts the number of animations in half.  The only difference is that
the keypad will be slightly brighter in mid-fade.  This can be explained
as follows:  Imagine that both colors were white.  With this change, you
would always see white, whereas before, in mid-fade you'd get half white
and half half-white, i.e. 75% light gray.  Arguably, the new behavior is
correct.